fmt + improve bundle.sh

+++ build/bundle/bundle.sh
@@ -26,11 +26,25 @@ mkdir /tmp/craftip-bundle
 
 unzip $1 -d $BUNDLE/input
 
-cargo run -- --input $BUNDLE/input/bin --output $BUNDLE/output --ver `cat ${BUNDLE}/input/version`
+cargo run -- --input $BUNDLE/input/bin --output $BUNDLE/output --ver $(cat "${BUNDLE}/input/version")
+echo "Uploading updater files into staging..."
 scp -r $BUNDLE/output/* "${CRAFTIP_SERVER}:${CRAFTIP_SERVER_DIR}/update/v1/"
+echo "Testing staging..."
 cargo run -- --test-staging
+echo "Uploading Binaries to be downloaded from the website"
 scp -r $BUNDLE/input/downloads/ "${CRAFTIP_SERVER}:${CRAFTIP_SERVER_DIR}/downloads_staging"
+echo "Moving everything from staging to production"
+
+read -r -p "Are you sure? [y/N] " response
+if [[ ! $response =~ ^([yY][eE][sS]|[yY])$ ]]
+then
+    exit 1
+fi
+
+
 ssh ${CRAFTIP_SERVER} "\
-    mv ${CRAFTIP_SERVER_DIR}/update/v1/latest.json.staging.json ${CRAFTIP_SERVER_DIR}/update/v1/latest.json;\
-    rm -rf ${CRAFTIP_SERVER_DIR}/downloads;\
-    mv ${CRAFTIP_SERVER_DIR}/downloads_staging ${CRAFTIP_SERVER_DIR}/downloads"
+    mv ${CRAFTIP_SERVER_DIR}/update/v1/latest.json.staging.json ${CRAFTIP_SERVER_DIR}/update/v1/latest.json;
+    mv ${CRAFTIP_SERVER_DIR}/downloads /tmp/downloads-$(date +%s);
+    mv ${CRAFTIP_SERVER_DIR}/downloads_staging ${CRAFTIP_SERVER_DIR}/downloads;"
+
+echo "Done!"
